+++ crates/px/src/commands/masters/create_flat.rs
@@ -1,6 +1,7 @@
 use crate::{ExitStatus, printer::Printer, utils::to_fits_ext};
 use anyhow::Result;
 use px_cli::CreateFlatMasterArgs;
+use px_conventions::profile::ProfilePath;
 use px_fits::{CalibrationMetadata, all_fits_files};
 use siril_sys::{
     Builder,
@@ -12,6 +13,7 @@ use std::path::PathBuf;
 pub(crate) async fn create_master_flat(
     args: CreateFlatMasterArgs,
     printer: Printer,
+    profile_path: ProfilePath,
 ) -> Result<ExitStatus> {
     // Guard to make sure the input folder exists first
     if !args.raw_folder.exists() {
@@ -26,11 +28,15 @@ pub(crate) async fn create_master_flat(
         return Ok(ExitStatus::Error);
     }
 
-    // Guard to make sure the output folder exists
-    if !args.out_folder.exists() {
-        printer.error("Output flat folder does not exist")?;
-        return Ok(ExitStatus::Error);
-    }
+    // Get the output folder from args OR profile convention
+    let out_folder = match args.out_folder {
+        Some(ref path) if path.exists() => path.clone(),
+        Some(_) => {
+            printer.error("Output flat folder does not exist")?;
+            return Ok(ExitStatus::Error);
+        }
+        None => profile_path.flat.clone(),
+    };
 
     // TODO: Make bias optional and find the best by default else use specified or error
 
@@ -42,7 +48,7 @@ pub(crate) async fn create_master_flat(
 
     // Setup the output file
     let name = CalibrationMetadata::from(raw_files.first().unwrap())?.master_flat_name(args.filter);
-    let output_file = args.out_folder.join(name).display().to_string();
+    let output_file = out_folder.join(name).display().to_string();
 
     // Setup siril
     let ext = to_fits_ext(args.ext);