+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Add `px project add` to add an observation to the project and start filling out linear stack creation
 * Add `px project stack` and basic implementation with a number of todo's
 * Add `px profile init` and a convention plus profile config file
+* Enforce profile path being specified on `px master ..` commands
 
 ## 0.1.3